Move client logic to internal
