+++ cmd/client/cmd/login.go
@@ -9,6 +9,7 @@ import (
 	"log"
 	"time"
 
+	"github.com/shipherman/speech-to-text/internal/client"
 	"github.com/spf13/cobra"
 )
 
@@ -18,7 +19,7 @@ var loginCmd = &cobra.Command{
 	Short: "Login to STT service",
 	Long:  `Login command returns JWT token to use it for *transcribe* subcommand`,
 	Run: func(cmd *cobra.Command, args []string) {
-		c, err := NewClient()
+		c, err := client.NewClient(cfg)
 		if err != nil {
 			log.Println(err)
 		}
