+++ cmd/server/cmd/root.go
@@ -16,8 +16,8 @@ import (
 	_ "github.com/lib/pq"
 	"github.com/shipherman/speech-to-text/gen/ent"
 	sttservice "github.com/shipherman/speech-to-text/gen/stt/service/v1"
-	"github.com/shipherman/speech-to-text/internal/clients"
 	"github.com/shipherman/speech-to-text/internal/db"
+	clients "github.com/shipherman/speech-to-text/internal/lib"
 	"github.com/shipherman/speech-to-text/internal/logger"
 	"github.com/shipherman/speech-to-text/internal/services/auth"
 	"github.com/shipherman/speech-to-text/pkg/fsstore"
