+++ internal/client/client_auth.go
@@ -1,4 +1,4 @@
-package cmd
+package client
 
 import (
 	"context"
