+++ internal/lib/stt.go
